#25: Generation of swagger.json / OAS
